better assertions
