+++ varpro/Cargo.toml
@@ -25,6 +25,7 @@ num-complex = "0.4"
 criterion = "0.4"
 pprof = {version = "0.11",features = ["criterion","flamegraph"]}
 shared_test_code = {path = "../shared_test_code"}
+assert_matches = "1.5"
 
 [[bench]]
 name = "double_exponential_without_noise"
